feat(#210) Moved zero-volume corrector to functional interface.
